Directories refactoring + add DI
