+++ ContactBook/.idea/misc.xml
@@ -16,7 +16,7 @@
         <entry key="..\:/University/AndroidApplications/ContactBookProject/ContactBook/app/src/main/res/drawable/separator.xml" value="0.403125" />
         <entry key="..\:/University/AndroidApplications/ContactBookProject/ContactBook/app/src/main/res/drawable/text_input_edit_text_authorization.xml" value="0.403125" />
         <entry key="..\:/University/AndroidApplications/ContactBookProject/ContactBook/app/src/main/res/layout/activity_authentication.xml" value="0.28020833333333334" />
-        <entry key="..\:/University/AndroidApplications/ContactBookProject/ContactBook/app/src/main/res/layout/activity_main.xml" value="0.33" />
+        <entry key="..\:/University/AndroidApplications/ContactBookProject/ContactBook/app/src/main/res/layout/activity_main.xml" value="0.2" />
         <entry key="..\:/University/AndroidApplications/ContactBookProject/ContactBook/app/src/main/res/layout/contact_row.xml" value="0.28020833333333334" />
         <entry key="..\:/University/AndroidApplications/ContactBookProject/ContactBook/app/src/main/res/layout/custom_row.xml" value="0.5" />
         <entry key="..\:/University/AndroidApplications/ContactBookProject/ContactBook/app/src/main/res/layout/fragment_add_contact.xml" value="0.28020833333333334" />
