+++ ContactBook/app/src/main/AndroidManifest.xml
@@ -3,6 +3,7 @@
     package="com.example.contactbook">
 
     <application
+        android:name=".di.HiltInit"
         android:allowBackup="true"
         android:icon="@mipmap/ic_launcher"
         android:label="@string/app_name"
@@ -10,15 +11,17 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.ContactBook">
         <activity
-            android:name=".AuthenticationActivity"
-            android:exported="true" />
-        <activity
-            android:name=".MainActivity"
-            android:exported="true">
+            android:name=".ui.activities.AuthenticationActivity"
+            android:exported="true" >
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
+        </activity>
+        <activity
+            android:name=".ui.activities.MainActivity"
+            android:exported="true">
+
         </activity>
     </application>
 </manifest>
