--- ContactBook/app/src/main/java/com/example/contactbook/AuthenticationActivity.kt
@@ -1,11 +1,0 @@
-package com.example.contactbook
-
-import androidx.appcompat.app.AppCompatActivity
-import android.os.Bundle
-
-class AuthenticationActivity : AppCompatActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_authentication)
-    }
-}
