--- ContactBook/app/src/main/java/com/example/contactbook/data/daos/IContactDao.kt
@@ -1,18 +1,0 @@
-package com.example.contactbook.data.daos
-
-import androidx.lifecycle.LiveData
-import androidx.room.Dao
-import androidx.room.Insert
-import androidx.room.OnConflictStrategy
-import androidx.room.Query
-import com.example.contactbook.data.entities.Contact
-
-@Dao
-interface IContactDao {
-
-    @Insert(onConflict = OnConflictStrategy.IGNORE)
-    suspend fun addContact(contact : Contact)
-
-    @Query("SELECT * FROM contactsTable WHERE ownerId = :ownerId")
-    fun getContacts(ownerId : String) : LiveData<List<Contact>>
-}
