+++ ContactBook/app/src/main/java/com/example/contactbook/data/services/abstractions/IAuthentificateService.kt
@@ -1,0 +1,7 @@
+package com.example.contactbook.data.services.abstractions
+
+interface IAuthentificateService {
+    fun isUserExists(email : String, password : String) : Boolean
+
+    fun loginUser(email : String, password: String)
+}
