+++ ContactBook/app/src/main/java/com/example/contactbook/data/services/abstractions/IHashService.kt
@@ -1,0 +1,8 @@
+package com.example.contactbook.data.services.abstractions
+
+interface IHashService {
+
+    fun getHash(inputString : String, algorithm : String): String
+
+    fun toHex(byteArray : ByteArray): String
+}
