+++ ContactBook/app/src/main/java/com/example/contactbook/di/HiltInit.kt
@@ -1,0 +1,7 @@
+package com.example.contactbook.di
+
+import android.app.Application
+import dagger.hilt.android.HiltAndroidApp
+
+@HiltAndroidApp
+class HiltInit : Application()
