--- ContactBook/app/src/main/java/com/example/contactbook/screens/Main/contacts/ContactDetailFragment.kt
@@ -1,23 +1,0 @@
-package com.example.contactbook.screens.Main.contacts
-
-import android.os.Bundle
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import androidx.fragment.app.Fragment
-import com.example.contactbook.databinding.FragmentContactDetailBinding
-
-class ContactDetailFragment : Fragment() {
-    private var _binding: FragmentContactDetailBinding? = null
-    private val binding get() = _binding!!
-    override fun onCreateView(
-        inflater: LayoutInflater, container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View? {
-        _binding = FragmentContactDetailBinding.inflate(inflater, container, false)
-
-        return binding.root
-    }
-
-
-}
