--- ContactBook/app/src/main/java/com/example/contactbook/services/HashService.kt
@@ -1,17 +1,0 @@
-package com.example.contactbook.services
-
-import com.example.contactbook.services.abstractions.IHashService
-import java.security.MessageDigest
-
-class HashService : IHashService {
-
-    override fun getHash(inputString : String, algorithm : String): String{
-        val bytes = MessageDigest.getInstance(algorithm).digest(inputString.toByteArray())
-        return toHex(bytes)
-    }
-
-    override fun toHex(byteArray : ByteArray): String {
-        return byteArray.joinToString (""){ "%02x".format(it) }
-    }
-
-}
