--- ContactBook/app/src/main/java/com/example/contactbook/services/abstractions/IAuthentificateService.kt
@@ -1,7 +1,0 @@
-package com.example.contactbook.services.abstractions
-
-interface IAuthentificateService {
-    fun isUserExists(email : String, password : String) : Boolean
-
-    fun loginUser(email : String, password: String)
-}
