--- ContactBook/app/src/main/java/com/example/contactbook/services/abstractions/IAuthorizedUserSharedPreferencesService.kt
@@ -1,13 +1,0 @@
-package com.example.contactbook.services.abstractions
-
-import com.example.contactbook.data.Model.UserModel
-
-interface IAuthorizedUserSharedPreferencesService {
-    fun saveCurrentUserData(user : UserModel)
-
-    fun deleteCurrentUserData()
-
-    fun loadCurrentUser() : UserModel
-
-    fun isAuthorized() : Boolean
-}
