--- ContactBook/app/src/main/java/com/example/contactbook/services/abstractions/IHashService.kt
@@ -1,8 +1,0 @@
-package com.example.contactbook.services.abstractions
-
-interface IHashService {
-
-    fun getHash(inputString : String, algorithm : String): String
-
-    fun toHex(byteArray : ByteArray): String
-}
