+++ ContactBook/app/src/main/java/com/example/contactbook/ui/activities/AuthenticationActivity.kt
@@ -1,0 +1,14 @@
+package com.example.contactbook.ui.activities
+
+import androidx.appcompat.app.AppCompatActivity
+import android.os.Bundle
+import com.example.contactbook.R
+import dagger.hilt.android.AndroidEntryPoint
+
+@AndroidEntryPoint
+class AuthenticationActivity : AppCompatActivity() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(R.layout.activity_authentication)
+    }
+}
