+++ ContactBook/app/src/main/java/com/example/contactbook/ui/views/Main/settings/SettingsFragment.kt
@@ -1,0 +1,23 @@
+package com.example.contactbook.ui.views.Main.settings
+
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.fragment.app.Fragment
+import com.example.contactbook.databinding.FragmentSettingsBinding
+
+class SettingsFragment : Fragment() {
+    private var _binding: FragmentSettingsBinding? = null
+    private val binding get() = _binding!!
+
+
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        _binding = FragmentSettingsBinding.inflate(inflater, container, false)
+
+        return binding.root
+    }
+}
