+++ ContactBook/build.gradle
@@ -7,7 +7,7 @@ buildscript {
     dependencies {
         classpath "com.android.tools.build:gradle:7.0.2"
         classpath 'org.jetbrains.kotlin:kotlin-gradle-plugin:1.5.31'
-
+        classpath 'com.google.dagger:hilt-android-gradle-plugin:2.40.1'
         // NOTE: Do not place your application dependencies here; they belong
         // in the individual module build.gradle files
     }
